python/requirements.txt
flask==2.2.2
mysql-connector-python==8.0.26
prometheus-client==0.11.0
Werkzeug==2.2.2